+++ src/compounds/billboard/src/index.tsx
@@ -73,17 +73,15 @@ const Billboard: React.FC<Props> = ({
                   ? ['xl', 'xxl', '0']
                   : ['xl', 'xxl'],
                 display: 'flex',
-                height: [
-                  'auto',
-                  'auto',
-                  !primarySlot && bgImage ? '402px' : 'auto'
-                ],
+                height: ['auto', 'auto', 'auto'],
                 flexDirection: [
                   'column' as const,
                   'column' as const,
                   'row' as const
                 ],
-                justifyContent: ['center', 'center', 'space-between']
+                justifyContent: ['center', 'center', 'space-between'],
+                position: 'relative',
+                alignItems: 'flex-end'
               }}
             >
               <div
@@ -95,7 +93,8 @@ const Billboard: React.FC<Props> = ({
                       : ('center' as const),
                   variant: styles(),
                   marginX: primarySlot || bgImage ? ['0', 'auto', '0'] : 'auto',
-                  width: ['auto', '70%', '50%']
+                  width: ['auto', '70%', '50%'],
+                  '& > ul > li': { fontSize: ['sm', 'md'], paddingY: 10 }
                 }}
               >
                 {primaryContent}
@@ -120,8 +119,6 @@ const Billboard: React.FC<Props> = ({
                     width: '404px',
                     height: '416px',
                     display: ['none', 'none', 'block'],
-                    position: 'relative' as const,
-                    top: '-14px',
                     background: `url(${bgImage}) no-repeat right bottom / contain`
                   }}
                 ></div>